// src/generated/api/model/planResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Fluenza API
 * OpenAPI spec version: 0.0.1
 */

export interface PlanResponse {
  priceId: string;
  unitAmount: number;
  maxKeys: number;
  rateLimitRpd: number;
  maxDomains: number;
  rateLimitRpm: number;
  /** @nullable */
  interval?: string | null;
  productName: string;
  currency: string;
}